pyproject.toml
[tool.poetry]
name = "swift-demo"
version = "0.0.1"
description = ""
authors = ["Kyle Getrost <kyle@motionhungry.com>"]

[tool.poetry.dependencies]
python = "^3.10"
load-dotenv = "^0.1.0"
requests = "^2.28.2"
PyJWT = "^2.6.0"
cryptography = "^39.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
